rubble when you destroy enemies. Particles that don't need physics now don't have them. MachineGun weapon. Attached machine gun weapon to bigBoy. New release. Tuning.
git-svn-id: http://www.piratestudios.com/pammo/svn@281 fc494e2b-ab4b-0410-a0f4-e0a3d7fb9d21

+++ client/Source/Engine2/enemyLoader.cpp
@@ -218,7 +218,7 @@ bool EnemyLoader::parseWeaponMachineGun(char* s)
             data->mMaxDistance = atol(s);
             break;
         case 12: // Firing rate
-            data->mFireRate = atol(s);
+            data->mFireRate = (float)atof(s);
             ++mTemplate->mWeaponCount;
             return true;
         }
